requirements.txt
python-dotenv
langchain-huggingface
pdfplumber
python-docx
openpyxl
python-pptx
streamlit